python changes.
Reorganized python package.
Added python .gitignore entry.
Added some new commands.
