+++ .gitignore
@@ -4,3 +4,5 @@ hail.iml
 .gradle/
 build/
 out/
+python/**/*.pyc
+.*.crc
